requirements.txt
# Core Libraries
pandas==2.1.1
numpy==1.26.3  # updated to meet jobspy requirement
scikit-learn==1.3.0
streamlit==1.27.0

# Natural Language Processing
nltk==3.8.1
transformers==4.34.0
torch==2.2.1

# PDF Processing
PyPDF2==3.0.1

# Job Data Fetching
python-jobspy==1.1.78  # Requires Python >=3.10, <4.0

# Keyword Extraction
rake-nltk==1.0.6
